Add network_utils directory
