+++ fastpay/Cargo.toml
@@ -11,7 +11,6 @@ anyhow = "1.0.53"
 bytes = "1.1.0"
 clap = "3.0.13"
 futures = "0.3.19"
-net2 = "0.2.37"
 serde = { version = "1.0.136", features = ["derive"] }
 serde_json = "1.0.78"
 structopt = "0.3.26"
@@ -34,6 +33,7 @@ tracing-subscriber = { version = "0.3", features = ["time", "env-filter"] }
 bcs = "0.1.3"
 fastpay_core = { path = "../fastpay_core" }
 fastx-adapter = { path = "../fastx_programmability/adapter" }
+fastx-network = { path = "../network_utils" }
 fastx-types = { path = "../fastx_types" }
 
 move-package = { git = "https://github.com/diem/move", rev="62b5a5075378ae6a7102bbfc1fb33b57ba6125d2" }
@@ -51,6 +51,3 @@ path = "src/server.rs"
 [[bin]]
 name = "bench"
 path = "src/bench.rs"
-
-[package.metadata.cargo-udeps.ignore]
-normal = ["net2"]
