+++ fastpay/src/server.rs
@@ -3,8 +3,9 @@
 
 #![deny(warnings)]
 
-use fastpay::{config::*, network, transport};
+use fastpay::{config::*, server_lib};
 use fastpay_core::authority::*;
+use fastx_network::transport;
 use fastx_types::{base_types::*, committee::Committee, object::Object};
 
 use futures::future::join_all;
@@ -23,7 +24,7 @@ fn make_server(
     committee_config_path: &str,
     initial_accounts_config_path: &str,
     buffer_size: usize,
-) -> network::Server {
+) -> server_lib::Server {
     let server_config =
         AuthorityServerConfig::read(server_config_path).expect("Fail to read server config");
     let committee_config =
@@ -61,7 +62,7 @@ fn make_server(
         state
     });
 
-    network::Server::new(
+    server_lib::Server::new(
         local_ip_addr.to_string(),
         server_config.authority.base_port,
         state,
