+++ fastpay_core/src/client.rs
@@ -4,8 +4,9 @@
 use crate::downloader::*;
 use async_trait::async_trait;
 use fastx_framework::build_move_package_to_bytes;
+use fastx_network::network;
 use fastx_types::{
-    base_types::*, committee::Committee, error::FastPayError, fp_ensure, messages::*,
+    base_types::*, committee::Committee, error::FastPayError, fp_ensure, messages::*, serialize::*,
 };
 use futures::{future, StreamExt, TryFutureExt};
 use move_core_types::identifier::Identifier;
@@ -58,6 +59,46 @@ pub trait AuthorityClient {
         &self,
         request: ObjectInfoRequest,
     ) -> Result<ObjectInfoResponse, FastPayError>;
+}
+
+#[async_trait]
+impl AuthorityClient for network::Client {
+    /// Initiate a new transfer to a FastPay or Primary account.
+    async fn handle_order(&mut self, order: Order) -> Result<OrderInfoResponse, FastPayError> {
+        self.send_recv_bytes(serialize_order(&order), order_info_deserializer)
+            .await
+    }
+
+    /// Confirm a transfer to a FastPay or Primary account.
+    async fn handle_confirmation_order(
+        &mut self,
+        order: ConfirmationOrder,
+    ) -> Result<OrderInfoResponse, FastPayError> {
+        self.send_recv_bytes(serialize_cert(&order.certificate), order_info_deserializer)
+            .await
+    }
+
+    async fn handle_account_info_request(
+        &self,
+        request: AccountInfoRequest,
+    ) -> Result<AccountInfoResponse, FastPayError> {
+        self.send_recv_bytes(
+            serialize_account_info_request(&request),
+            account_info_deserializer,
+        )
+        .await
+    }
+
+    async fn handle_object_info_request(
+        &self,
+        request: ObjectInfoRequest,
+    ) -> Result<ObjectInfoResponse, FastPayError> {
+        self.send_recv_bytes(
+            serialize_object_info_request(&request),
+            object_info_deserializer,
+        )
+        .await
+    }
 }
 
 pub struct ClientState<AuthorityClient> {
